[XHAB-38] renaming files to resolve conflicts
