+++ lib/src/domain/splash_screen_use_case.dart
